refactor getting remainingTeamSlots
